refactor: create a shared ui segment
