+++ src/features/game/card/PaintingInfo.ts
@@ -1,6 +1,5 @@
-import Component from "../../../shared/Component";
+import { Component, p } from "../../../shared/ui";
 import { Observer, Publisher } from "../../../shared/Observer";
-import { p } from "../../../ui/tags";
 import RoundState, { type Painting } from "../model/RoundState";
 
 import styles from "./PaintingInfo.module.css";
