+++ src/features/game/fields/Row.ts
@@ -1,7 +1,6 @@
-import Component from "../../../shared/Component";
+import { Component, div } from "../../../shared/ui";
 import { debounceListener } from "../../../shared/helpers";
 import { Observer, Publisher } from "../../../shared/Observer";
-import { div } from "../../../ui/tags";
 
 import WordCard, { type Word } from "../card/WordCard";
