+++ src/features/game/stats/ArtworkMiniature.ts
@@ -1,5 +1,4 @@
-import Component from "../../../shared/Component";
-import { div, img, p } from "../../../ui/tags";
+import { Component, div, img, p } from "../../../shared/ui";
 
 import { type Painting } from "../model/RoundState";
 import { IMAGES_BASE_URL } from "../../../shared/helpers";
