+++ src/features/game/stats/SentencesList.ts
@@ -1,6 +1,12 @@
-import Component from "../../../shared/Component";
-import AudioButton from "../../../ui/button/AudioButton";
-import { h3, li, p, span, ul } from "../../../ui/tags";
+import {
+  Component,
+  AudioButton,
+  h3,
+  li,
+  p,
+  span,
+  ul,
+} from "../../../shared/ui";
 import { type Stage } from "../model/RoundState";
 
 import styles from "./SentencesList.module.css";
