+++ src/shared/ui/base/Component.ts
