+++ src/shared/ui/base/tags.ts
@@ -1,4 +1,4 @@
-import Component, { ElementFuncProps } from "../shared/Component";
+import Component, { ElementFuncProps } from "./Component";
 
 export const div = (
   props: ElementFuncProps<HTMLDivElement>,
