+++ src/shared/ui/button/Button.ts
@@ -1,4 +1,4 @@
-import Component from "../../shared/Component";
+import Component from "../base/Component";
 
 import styles from "./Button.module.css";
