+++ src/shared/ui/index.ts
@@ -1,0 +1,14 @@
+export {
+  default as Component,
+  type ElementFuncProps,
+  type Props,
+} from "./base/Component";
+export * from "./base/tags";
+export { default as Button } from "./button/Button";
+export { default as AudioButton } from "./button/AudioButton";
+export { default as ButtonIcon } from "./button/ButtonIcon";
+export { default as Form, type TextInputParams } from "./form/Form";
+export { default as Header } from "./header/Header";
+export { default as Input } from "./input/Input";
+export { default as Message, MessageType } from "./message/Message";
+export { default as Modal } from "./modal/Modal";
